% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Batch}
\alias{Batch}
\title{Example Batch Vector}
\format{
A vector with 100 elements, representing the batch 100 individuals came from (1 or 2).
}
\usage{
data(Batch)
}
\description{
This is an example of a vector including batch information.
}
\examples{
data(Batch)
}
\keyword{datasets}
