% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_combat.R
\name{cam_combat}
\alias{cam_combat}
\title{Apply CAM ComBat}
\usage{
cam_combat(Y, Batch, X, res = FALSE)
}
\arguments{
\item{Y}{(p x n) data matrix for which the p rows are features, and the n columns are participants}

\item{Batch}{(n x 1) numeric or character vector of length n indicating the site/scanner/study id}

\item{X}{(n x b) matrix, with n as the number of individuals, and b as the number of covariates}

\item{res}{option to determine if covariates should be included or not (default is FALSE)}
}
\value{
Returns list in which the last element is the final harmonized dataset and the other elements are the batch-specific covariance estimates
}
\description{
This function applies Covariance-Aware Multivariate (CAM) ComBat to harmonize a given dataset.
}
\examples{
data(Y)
data(X)
data(Batch)
Youtput3 <- cam_combat(Y,Batch,X)

}
