% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_combat.R
\name{sib_combat}
\alias{sib_combat}
\title{Apply SIB ComBat}
\usage{
sib_combat(
  Y,
  Batch,
  X,
  nclust = NULL,
  Clust_list = NULL,
  niter,
  ncores,
  res = FALSE,
  pres = FALSE,
  cb = FALSE
)
}
\arguments{
\item{Y}{(p x n) data matrix for which the p rows are features, and the n columns are participants}

\item{Batch}{(n x 1) numeric or character vector of length n indicating the site/scanner/study id}

\item{X}{(n x b) matrix, with n as the number of individuals, and b as the number of covariates}

\item{nclust}{number of clusters to divide features in if no assignments are inputted}

\item{Clust_list}{list of cluster assignments for features for each SIB ComBat iteration (if NULL, assignments are determined randomly)}

\item{niter}{number of SIB ComBat iterations to perform}

\item{ncores}{number of cores to parallelize SIB ComBat harmonization across}

\item{res}{option to determine if covariates should be included or not (default is FALSE)}

\item{pres}{option to determine if jobs should be assigned dynamically when parallelizing (default is FALSE)}

\item{cb}{option to determine if ComBat should be applied first (default is FALSE)}
}
\value{
Returns final harmonized data after performing SIB ComBat
}
\description{
This function applies Spatially-Informed Iterative Block (SIB) ComBat to harmonize a given dataset.
}
\examples{
data(Y)
data(X)
data(Batch)
Yharm4 <- sib_combat(Y,Batch,X,nclust=2,Clust_list=NULL,niter=2,ncores=1,res=FALSE,pres=FALSE,cb=FALSE)

}
