% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{Example Covariate Vector}
\format{
A vector with 100 elements, representing values of a binary covariate (0 or 1) from 100 individuals.
}
\usage{
data(X)
}
\description{
This is an example of a vector (or matrix) including covariate information.
}
\examples{
data(X)
}
\keyword{datasets}
