% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{Example Matrix of Imaging Outcomes}
\format{
A matrix with 100 rows and 100 columns, corresponding to 100 individuals each with 100 features.
}
\usage{
data(Y)
}
\description{
This is an example of a matrix of imaging outcomes.
}
\examples{
data(Y)
}
\keyword{datasets}
